# Benchmarks for timeseries `aggregate_downsample()`