Would have caught the performance regression in v7.1.0 ( https://github.com/astropy/astropy/issues/18176 ), being fixed by PR https://github.com/astropy/astropy/pull/18188 .

The benchmark results on my machine

| `aggregate_func` | v7.1.0     | with astropy PR #18188 |
| ---------------- | ---------- | ---------------------- |
| `None`           | 9.49±0.2ms | 9.79±0.3ms             |
| `np.nanmean`     | 186±20ms   | 119±3ms                |


@mvhk @pllim