@orionlee 2025-06-03
The issues raised by mvhk have been resolved.

The test is now more granular, one for each column type - `aggregate_func` combination

```txt
With PR 18188:
              --                           aggregate_func
              ---------- --------------------------------------------------
               col_type       None      <function nanmean>   <ufunc 'add'>
              ========== ============= ==================== ===============
                 col       7.21±0.3ms       24.5±0.4ms          7.03±4ms
                 mcol     7.28±0.08ms        69.4±3ms          7.32±0.6ms
                 qty       6.97±0.6ms       11.8±0.3ms        6.89±0.04ms
                 mqty     7.35±0.07ms       28.2±0.2ms          217±4ns
              ========== ============= ==================== ===============

Before the PR (v7.1.0-like)

              --                           aggregate_func
              ---------- -------------------------------------------------
               col_type      None      <function nanmean>   <ufunc 'add'>
              ========== ============ ==================== ===============
                 col      7.00±0.2ms       24.6±0.4ms         6.95±0.1ms
                 mcol      7.23±2ms        69.8±10ms          7.31±0.1ms
                 qty      6.99±0.1ms        31.5±8ms          6.89±0.1ms
                 mqty     7.25±0.1ms        59.3±9ms           214±70ns
              ========== ============ ==================== ===============
```

- It shows astropy PR 18188 addressed the performance regression in v7.1.0 for `np.nanmean` with `Quantity` / `MaskedQuantity`.
- The test also makes it explicit that for `np.nanmean` with `Column` / `MaskedColumn`, the performance is not good either. But I think it's an existing issue that has never been raised / addressed.